Move edit function into separate module and remove utils
